+++ packages/terra-form-select/tests/jest/FrameUtil.test.jsx
@@ -215,13 +215,13 @@ describe('FrameUtil', () => {
     it('should return false for a combobox variant if the search value is empty', () => {
       const props = { variant: Variants.COMBOBOX, onSelect: 'Mock', value: ['Tag'] };
       const state = { hasSearchChanged: true, searchValue: '' };
-      expect(FrameUtil.shouldAddOptionOnBlur(props, state)).toBeTruthy();
+      expect(FrameUtil.shouldAddOptionOnBlur(props, state)).toBeFalsy();
     });
 
-    it('should return true for a combobox variant if the search value is blank', () => {
+    it('should return false for a combobox variant if the search value is blank', () => {
       const props = { variant: Variants.COMBOBOX, onSelect: 'Mock', value: ['Tag'] };
       const state = { hasSearchChanged: true, searchValue: '  ' };
-      expect(FrameUtil.shouldAddOptionOnBlur(props, state)).toBeTruthy();
+      expect(FrameUtil.shouldAddOptionOnBlur(props, state)).toBeFalsy();
     });
 
     it('should return false for a combobox variant if the search value has not changed', () => {